+++ lib/presentation/screens/movies/home_screen.dart
@@ -1,102 +1,17 @@
-import 'package:cinemapedia_app/presentation/providers/providers.dart';
-import 'package:cinemapedia_app/presentation/widgets/shared/full_screen_loader.dart';
 import 'package:cinemapedia_app/presentation/widgets/widgets.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_riverpod/flutter_riverpod.dart';
 
 // Pantalla principal de la aplicación
 class HomeScreen extends StatelessWidget {
   static const name = 'home-screen'; // Nombre para la navegación
-  const HomeScreen({super.key});
+  final Widget childView;
+  const HomeScreen({super.key, required this.childView});
 
   @override
   Widget build(BuildContext context) {
-    return const Scaffold(
-      body: _HomeView(),
+    return Scaffold(
+      body: childView,
       bottomNavigationBar: CustomBottomNavigationBar(),
     ); // Rendering la vista principal
-  }
-}
-
-// Widget que maneja el estado con Riverpod
-class _HomeView extends ConsumerStatefulWidget {
-  const _HomeView();
-
-  @override
-  _HomeViewState createState() => _HomeViewState();
-}
-
-class _HomeViewState extends ConsumerState<_HomeView> {
-  @override
-  void initState() {
-    super.initState();
-    // Cargar la primera página de películas en cartelera al iniciar
-
-    ref.read(nowPlayingMoviesProvider.notifier).loadNextPage();
-    ref.read(popularMoviesProvider.notifier).loadNextPage();
-    ref.read(upComingMoviesProvider.notifier).loadNextPage();
-    ref.read(toRatedMoviesProvider.notifier).loadNextPage();
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    final initialLoading = ref.watch(initialLoadingProvider);
-    if (initialLoading) return const FullScreenLoader();
-    final nowPlayingMovies = ref.watch(nowPlayingMoviesProvider);
-    final slideShowMovies = ref.watch(moviesSlideShowProvider);
-    final popularMovies = ref.watch(popularMoviesProvider);
-    final upComingMovies = ref.watch(upComingMoviesProvider);
-    final toRatedMovies = ref.watch(toRatedMoviesProvider);
-
-    return CustomScrollView(
-      slivers: [
-        const SliverAppBar(
-          floating: true,
-          flexibleSpace: FlexibleSpaceBar(title: CustomAppBar()),
-        ),
-        SliverList(
-          delegate: SliverChildBuilderDelegate((context, index) {
-            return Column(
-              children: [
-                MoviesSlideShow(movies: slideShowMovies),
-                MoviesHorizontalListView(
-                  movies: nowPlayingMovies,
-                  title: 'En cines',
-                  subTitle: 'Lunes 20',
-                  loadNextPage: () {
-                    ref.read(nowPlayingMoviesProvider.notifier).loadNextPage();
-                  },
-                ),
-                MoviesHorizontalListView(
-                  movies: upComingMovies,
-                  title: 'Próximamente',
-                  subTitle: 'En este mes',
-                  loadNextPage: () {
-                    ref.read(upComingMoviesProvider.notifier).loadNextPage();
-                  },
-                ),
-                MoviesHorizontalListView(
-                  movies: popularMovies,
-                  title: 'Populares',
-                  //subTitle: 'En este mes',
-                  loadNextPage: () {
-                    ref.read(popularMoviesProvider.notifier).loadNextPage();
-                  },
-                ),
-                MoviesHorizontalListView(
-                  movies: toRatedMovies,
-                  title: 'Mejor calificadas',
-                  subTitle: 'Desde siempre',
-                  loadNextPage: () {
-                    ref.read(toRatedMoviesProvider.notifier).loadNextPage();
-                  },
-                ),
-                const SizedBox(height: 15),
-              ],
-            );
-          }, childCount: 1),
-        ),
-      ],
-    );
   }
 }
