+++ lib/presentation/widgets/shared/custom_bottom_navigation_bar.dart
@@ -1,12 +1,47 @@
 import 'package:flutter/material.dart';
+import 'package:go_router/go_router.dart';
 
 class CustomBottomNavigationBar extends StatelessWidget {
   const CustomBottomNavigationBar({super.key});
+
+  int getCurrentIndex(BuildContext context) {
+    final String location = GoRouterState.of(context).matchedLocation;
+
+    switch (location) {
+      case '/':
+        return 0;
+
+      case '/categories':
+        return 1;
+      case '/favorites':
+        return 2;
+
+      default:
+        return 0;
+    }
+  }
+
+  void onItemTapped(BuildContext context, int index) {
+    switch (index) {
+      case 0:
+        context.go('/');
+        break;
+      case 1:
+        context.go('/');
+        break;
+      case 2:
+        context.go('/favorites');
+        break;
+      default:
+    }
+  }
 
   @override
   Widget build(BuildContext context) {
     return BottomNavigationBar(
       elevation: 0,
+      currentIndex: getCurrentIndex(context),
+      onTap: (value) => onItemTapped(context, value),
       items: const [
         BottomNavigationBarItem(icon: Icon(Icons.home_max), label: 'Inicio'),
         BottomNavigationBarItem(
